Refactor JavaScript plugin
