+++ test/plugins/javascript.test.js
@@ -1,19 +1,19 @@
 import assert from 'assert';
-import javascriptUniversal from '../../lib/resolver/javascript-universal.js';
+import plugin from '../../lib/plugins/javascript.js';
 
 describe('javascript-universal', () => {
   it("resolves 'foo/bar.js' like 'foo'", () => {
     const type = 'npm';
     assert.deepEqual(
-      javascriptUniversal({ type, target: 'foo/bar.js' }),
-      javascriptUniversal({ type, target: 'foo' }),
+      plugin.resolve({ type, target: 'foo/bar.js' }),
+      plugin.resolve({ type, target: 'foo' }),
     );
   });
 
   it("resolves '@angular/core/bar.js' to '@angular/core'", () => {
     const type = 'npm';
     assert.deepEqual(
-      javascriptUniversal({ type, target: '@angular/core/bar.js' })[0],
+      plugin.resolve({ type, target: '@angular/core/bar.js' })[0],
       {
         url: 'https://githublinker.herokuapp.com/q/npm/@angular/core',
         method: 'GET',
@@ -23,14 +23,14 @@ describe('javascript-universal', () => {
 
   it("resolves 'module' to 'https://nodejs.org/api/modules.html'", () => {
     assert.deepEqual(
-      javascriptUniversal({ target: 'module' }),
+      plugin.resolve({ target: 'module' }),
       'https://nodejs.org/api/modules.html',
     );
   });
 
   it("resolves './modules/es6.symbol' without stripping .symbol suffix", () => {
     assert.deepEqual(
-      javascriptUniversal({ target: './modules/es6.symbol' }),
+      plugin.resolve({ target: './modules/es6.symbol' }),
       [
         '{BASE_URL}modules/es6.symbol.js',
         '{BASE_URL}modules/es6.symbol/index.js',
@@ -49,7 +49,7 @@ describe('javascript-universal', () => {
 
   it("resolves './modules/es6.symbol.js' like './modules/es6.symbol'", () => {
     assert.deepEqual(
-      javascriptUniversal({ target: './modules/es6.symbol.js' }),
+      plugin.resolve({ target: './modules/es6.symbol.js' }),
       [
         '{BASE_URL}modules/es6.symbol.js',
         '{BASE_URL}modules/es6.symbol/index.js',
